common: Move GTK event handlers to gui_events
