+++ common.h
@@ -104,12 +104,4 @@ guint    GetBin        (double Freq, guint FFTLen);
 void     *Listen       ();
 void     saveCurrentPic();
 
-void     evt_AbortRx       ();
-void     evt_changeDevices ();
-void     evt_clearPix      ();
-void     evt_clickimg      ();
-void     evt_deletewindow  ();
-void     evt_GetAdaptive   ();
-void     evt_ManualStart   ();
-
 #endif
