+++ gui_events.h
@@ -1,0 +1,12 @@
+#ifndef _GUI_EVENTS_H_
+#define _GUI_EVENTS_H_
+
+void     evt_AbortRx       ();
+void     evt_changeDevices ();
+void     evt_clearPix      ();
+void     evt_clickimg      ();
+void     evt_deletewindow  ();
+void     evt_GetAdaptive   ();
+void     evt_ManualStart   ();
+
+#endif
